# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    API for messaging app WASAText
  version: 0.0.1
security:
  - bearerAuth: [ ]
...
paths:
...
  /chats/{chatId}/image:
    parameters:
      - $ref: "#/components/parameters/ChatID"
    put:
      tags:
        - chats
      summary: Sets group profile photo
      description: "Updates the group profile picture with the one given in the request"
      operationId: setGroupPhoto
      requestBody:
        content:
          multipart/form-data:
            schema:
              description: group image
              type: string
              format: binary
              minLength: 1
              maxLength: 99999
        required: true
      responses:
        "200":
          description: Group image updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    description: successful operation
                    type: boolean
                    example: true
                  message:
                    description: response message
                    type: string
                    example: Group image updated successfully.
                  imageUrl:
                    description: image address
                    type: string
                    format: uri
                    example: https://cdn.example.com/users/12345/profile.jpg
        "400":
          $ref: '#/components/responses/BadRequestError'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "404":
          $ref: '#/components/responses/NotFoundError'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        message:
          type: string
          example: "The request could not be processed"
        code:
          type: integer
          example: 400
  parameters:
...
    ChatID:
      schema:
        type: number
        format: integer
        example: 134
      name: chatId
      in: path
      required: true
      description: Chat identifier
...
  responses:
...
    UnauthorizedError:
      description: Invalid or missing bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequestError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer